requirements/base.txt
dnspython==1.14.0
boto3==1.3.1
docopt==0.6.2